Fix compile for xuc.cc on Xcode12
The 1.6 release fails to complie on Xcode 12.1 because of "byte"
conflicting with "std::byte".  I noticed that this was already
renamed "byteval" in the master branch (part of the larger set
of changes in d7998ecace00daa0469494377f916db386396406).
I thought it would be useful to backport this to the v1.6.x
branch as well.