Verify we can link to freetype
The existing configure script just verifies that it can find the freetype2-config command in the path.  When cross-compiling, it mauy find the host's freetype2-config, but not actually have the libraries for the cross-compile target.  This patch adds an additional test to verify that the symbol FT_Init_FreeType can be resolved at link time by the toolchain.

Verified on a mac which had freetyp2-config in the path that a build with --enable-android-apk successfully enables freetype for the mac build, but disables it for the android sub-build.