Change Video defaults.
Upped the default resolution to 1024x768, in theory I think we could even go to 1200x960.
Changed the scaler default to Point as people are now more often asking for the "real" unfiltered experience.
And fill mode "Fit" as the "Centre" mode made the game window too small and confused first time users of Exult.