Add option to build Exult as a library
The existing build structure always tries to build exult as an executable.  This ends up causing link errors when building for android.  This patch introduces a configure option to build the entire exult codebase into a library instead of an executable.  This will enable android builds to opt out of the executable while still building the library which the app will load at runtime.  This configure option is disabled by default, so builds on other platforms will still generate a monolithic executable like they always have.