Fix ogg/vorbis stuttering on Android
The existing 20Hz buffering for ogg/vorbis playback isn't sufficient on android.  It results in stuttering when using digital music on both an emulated android device on a laptop and on two native android devices.  Increasing to 5Hz buffering.

Note: This value was chosen based on empirical testing on a Kindle Fire tablet.  6Hz and above resulted in stuttering with digital music enabled.