Upgrade SDL to 2.0.22
This patch updates the Exult Android build to use SDL v2.0.22 and switches from using `configure` to using `cmake` for configuration.  This change is partly just to stay updated to the latest SDL release, but also leverages the `CMakeLists.txt` provided with SDL which is preconfigured for Android rather than trying to manually set flags for the `configure` script to set it up correctly for Android.  With this change, the build now includes hidapi as well, which was previously being disabled due to link errors.