Fix inverted logic in Image_window::Get_best_bpp()`
This patch fixes some inverted logic in `Get_best_bpp()`, which appeared to be expecting `VideoModeOK()` to return 0 on success.  This was causing erroneous error messages in the logs ("surface is not OK") and causing scalers to fail on Android.  With this change, `Get_best_bpp()` now returns the highest bpp detected by `VideoModeOK()`.

Note that there are a number of places in `imagewin.cc` where `VideoModeOK()` appears to be used incorrectly/inconsistently because the implementation completely ignores the `bpp` and `flags` parameters.  This [merits further cleanup](https://github.com/exult/exult/issues/220), but this patch does not attempt to address the broader issues.