Use exit() to terminate NDK code
Simply returning from the NDK entrypoint function does not clean up global resources since the process has not terminated, which can cause problems with Exult as it assumes the entire engine terminates upon return from `main()`.  Calling `exit()` ensures full cleanup so that the Android version behaves more like a standard process exiting.  The need for this surfaced when debugging problems with enabling scalers on Android.  They work the first time the app launches, but fail on subsequent runs.  With this patch applied, scalers work consistently on every run.  This patch is a workaround which is technically masking real bugs with proper resource cleanup in the code, but as the entire Exult codebase is written to assume complete cleanup upon return from main, there may be many small problems to chase down and it seems safer to use `exit()` to ensure consistent behavior.