Remove unused parameters from Image_window::VideoModeOK()
This function was originally introduced to mimic SDL_VideoModeOK(), which was deprecated with SDL2.  That function accepted `bpp` and `flags` parameters, but this implementation has never used them, resulting in some misleading behavior in the calling code.  This patch simply removes the unused flags and cleans up the calling code to eliminate redundancy.  This patch should not result in any functional change.