Improving blood/slime trails
NPCs will now be set to bleed when hurt badly,
and this will happen over time.
Instead of dropping multiple blood pools in the
same tile, the size of the blood will increase
as the NPC bleeds on the same tile.
Also made slimes use the same code when they
move around, and increased the chance of them
dropping slime trails.