Correcting UI_is_water behavior
In the original, it ignores Z coordinate and always tests
the tile at the resulting (X, Y) location.
Added Exult extensions to pass an object, or a
(object, x, y, z) position to test if an object has the
'water' flag set. In the latter case, if object == null
(such as when clicking on a tile), it uses the (x, y)
position to inspect the tile.
Making use of new functionality on Keyring and
SI Fixes for the fishing rod for better detection of
places you can fish on.