Fixing data split among bundle/system data dir
Exult searches for exult.flx to determine a valid data path.
When it searches for for it in CWD/data, due to the way SDL_RWops
works on OSX, it would search again in the bundle, which
is not what we want. This leaves <BUNDLE> defined but <DATA> as
CWD/data, which does not use the EXULT_DATADIR.

The solution is to avoid checking for CWD/data when
we found the App bundle.

Fixes #328 macOS EXULT_DATADIR trouble (caused by Android changes)