Fixing music and call opcodes
It seems that the originals deal with bytes internally for scripts,
whereas Exult deals in usecode values. This is noticeable
with call opcodes when function is < 0x100: there is always
another zero usecode value after the call in these cases.

For music opcode: there is a second parameter for continuous
flag; having a (nonzero)high byte in the song works.

This likely means our implementation of 'repeat'and 'repeat2'is wrong,
as it deals with usecode values. This is noticeable in a few places
in original usecode that have an off-by-one jump distance when
compiled by UCC.

No idea how to fix this, though.