Making UI_update_last_created work as in original
Exult was fixing a usecode bug in SI by changing the return value to
an object pointer, but this is wrong; it works the same as in BG.

Instead, I am fixing the intrinsic behavior and documenting the
usecode bug, with a proper fix.

Also fixing UI_update_last_created/give_last_created documentation.