Update Windows Badge in README.md
Updating Windows snapshot badge. Removed Windows label and icon because the automated snapshot are for both Windows and Android Changed the badge to actually show the info about the pre-release releases instead of only the latest official release. Changed the link to be point to a search of the pre-release builds as the latest release url doesn't point to pre release snapshots