Made ModManager::get_mod return ModInfo*
Simplifies a lot of things because these dynamic_casts are never needed.
Also fixed a potential null dereference warning that never actually happens.
Cleaned up how the mod title is stripped of \r and \r to turn all sequences of \r, \n, \t, and space to a single space (that is, avoid double spaces on Windows).