Removing hard-coded cap of 2048 flags
Now we can have up to 32768 flags, mostly due to usecode limitations. This could be changed at a later date.

Globals will be written in "pages" of 2048 flags, and compacted to the least number of pages sufficient for the last set flag.

Fixed a bug on Usecode machine where an invalid flag number (too low or too high) would result in a value not being popped from the Usecode stack -- there seem to be many other such issues.

Fixes #813 Remove hardcoded cap on flags